Revert "Feature/am 30 cp graph time between successes"
