+++ package.json
@@ -23,7 +23,6 @@
     "@capacitor/filesystem": "^1.0.0",
     "@capacitor/storage": "^1.0.0",
     "@ionic/angular": "^5.5.2",
-    "chart.js": "^3.3.2",
     "rxjs": "~6.6.0",
     "tslib": "^2.0.0",
     "zone.js": "~0.11.4"
