--- src/app/results/results-graphs/results-graphs.page.html
@@ -1,33 +1,0 @@
-
-  <ion-header [translucent]="true">
-    <ion-toolbar>
-      <ion-title>
-        Resultados de la ejercitacion - Hay uno repetido
-      </ion-title>
-    </ion-toolbar>
-  </ion-header>
-  
-  <ion-content>
-    <div class="ion-padding">
-      <ion-card>
-        <ion-card-header>
-          Desempeño cumulativo
-        </ion-card-header>
-        <ion-card-content>
-          <canvas #lineCumulativeCanvas style="position: relative; height:20vh; width:40vw"></canvas>
-        </ion-card-content>
-      </ion-card>
-    </div>
-  
-    <div class="ion-padding">
-      <ion-card>
-        <ion-card-header>
-          Tiempo entre aciertos - Partida N° 6
-        </ion-card-header>
-        <ion-card-content>
-          <canvas #lineTimeBetweenSuccessesCanvas style="position: relative; height:20vh; width:40vw"></canvas>
-        </ion-card-content>
-      </ion-card>
-    </div>
-  </ion-content>
-
