--- src/app/results/results-graphs/results-graphs.page.spec.ts
@@ -1,24 +1,0 @@
-import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
-import { IonicModule } from '@ionic/angular';
-
-import { ResultsGraphsPage } from './results-graphs.page';
-
-describe('ResultsGraphsPage', () => {
-  let component: ResultsGraphsPage;
-  let fixture: ComponentFixture<ResultsGraphsPage>;
-
-  beforeEach(waitForAsync(() => {
-    TestBed.configureTestingModule({
-      declarations: [ ResultsGraphsPage ],
-      imports: [IonicModule.forRoot()]
-    }).compileComponents();
-
-    fixture = TestBed.createComponent(ResultsGraphsPage);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  }));
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
