+++ src/app/results/results.page.html
@@ -8,8 +8,7 @@
   <ion-card *ngFor="let result of results">
     <ion-card-header>
       <ion-card-subtitle>{{result.date}}</ion-card-subtitle>
-      <ion-card-title>Jorgelina Cordero</ion-card-title>
-<!--       <ion-card-title>{{result.patient}}</ion-card-title> -->
+      <ion-card-title>{{result.patient}}</ion-card-title>
     </ion-card-header>
     <ion-card-content>
       <ion-chip *ngIf="result.canceled==false" color="success">
